TFP-3468 bytt til tl-basert endringsvurdering
