--- domenetjenester/behandling-revurdering/src/main/java/no/nav/foreldrepenger/behandling/revurdering/felles/ErEndringIUttak.java
@@ -1,7 +1,0 @@
-package no.nav.foreldrepenger.behandling.revurdering.felles;
-
-public interface ErEndringIUttak {
-
-    boolean vurder(UttakResultatHolder uttakresultatRevurderingOpt, UttakResultatHolder uttakresultatOriginalOpt);
-
-}
