--- domenetjenester/behandling-revurdering/src/main/java/no/nav/foreldrepenger/behandling/revurdering/ytelse/svp/ErEndringIUttakImpl.java
@@ -1,19 +1,0 @@
-package no.nav.foreldrepenger.behandling.revurdering.ytelse.svp;
-
-import javax.enterprise.context.ApplicationScoped;
-
-import no.nav.foreldrepenger.behandling.revurdering.felles.ErEndringIUttak;
-import no.nav.foreldrepenger.behandling.revurdering.felles.UttakResultatHolder;
-import no.nav.foreldrepenger.behandlingskontroll.FagsakYtelseTypeRef;
-
-@ApplicationScoped
-@FagsakYtelseTypeRef("SVP")
-public class ErEndringIUttakImpl implements ErEndringIUttak {
-
-    @Override
-    public boolean vurder(UttakResultatHolder uttakresultatRevurdering, UttakResultatHolder uttakresultatOriginal) {
-        return uttakresultatOriginal.vurderOmErEndringIUttak(uttakresultatRevurdering);
-    }
-
-}
-
