+++ domenetjenester/behandling-revurdering/src/test/java/no/nav/foreldrepenger/behandling/revurdering/ytelse/svp/UttakResultatHolderImplTest.java
@@ -22,7 +22,7 @@ public class UttakResultatHolderImplTest {
     @Test
     public void skal_returnere_false_når_det_ikke_er_noe_uttaksresultat() {
         // Arrange
-        UttakResultatHolderImpl uttakResultatHolder = new UttakResultatHolderImpl(Optional.empty());
+        UttakResultatHolderSVP uttakResultatHolder = new UttakResultatHolderSVP(Optional.empty());
 
         // Act
         boolean resultat = uttakResultatHolder.kontrollerErSisteUttakAvslåttMedÅrsak();
@@ -45,7 +45,7 @@ public class UttakResultatHolderImplTest {
             .medUttakResultatArbeidsforhold(arbeidsforhold1)
             .medUttakResultatArbeidsforhold(arbeidsforhold2)
             .build();
-        UttakResultatHolderImpl uttakResultatHolder = new UttakResultatHolderImpl(Optional.of(uttakResultat));
+        UttakResultatHolderSVP uttakResultatHolder = new UttakResultatHolderSVP(Optional.of(uttakResultat));
 
         // Act
         boolean resultat = uttakResultatHolder.kontrollerErSisteUttakAvslåttMedÅrsak();
@@ -72,7 +72,7 @@ public class UttakResultatHolderImplTest {
             .medUttakResultatArbeidsforhold(arbeidsforhold1)
             .medUttakResultatArbeidsforhold(arbeidsforhold2)
             .build();
-        UttakResultatHolderImpl uttakResultatHolder = new UttakResultatHolderImpl(Optional.of(uttakResultat));
+        UttakResultatHolderSVP uttakResultatHolder = new UttakResultatHolderSVP(Optional.of(uttakResultat));
 
         // Act
         boolean resultat = uttakResultatHolder.kontrollerErSisteUttakAvslåttMedÅrsak();
@@ -99,7 +99,7 @@ public class UttakResultatHolderImplTest {
             .medUttakResultatArbeidsforhold(arbeidsforhold1)
             .medUttakResultatArbeidsforhold(arbeidsforhold2)
             .build();
-        UttakResultatHolderImpl uttakResultatHolder = new UttakResultatHolderImpl(Optional.of(uttakResultat));
+        UttakResultatHolderSVP uttakResultatHolder = new UttakResultatHolderSVP(Optional.of(uttakResultat));
 
         // Act
         boolean resultat = uttakResultatHolder.kontrollerErSisteUttakAvslåttMedÅrsak();
